+++ jib-core/src/integration-test/java/com/google/cloud/tools/jib/builder/BuildImageStepsIntegrationTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Google Inc.
+ * Copyright 2018 Google LLC. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -17,8 +17,10 @@
 package com.google.cloud.tools.jib.builder;
 
 import com.google.cloud.tools.jib.Command;
+import com.google.cloud.tools.jib.cache.Caches;
 import com.google.cloud.tools.jib.image.ImageReference;
 import com.google.cloud.tools.jib.registry.LocalRegistry;
+import java.nio.file.Path;
 import org.junit.Assert;
 import org.junit.ClassRule;
 import org.junit.Rule;
@@ -44,11 +46,12 @@ public class BuildImageStepsIntegrationTest {
             .setMainClass("HelloWorld")
             .build();
 
+    Path cacheDirectory = temporaryCacheDirectory.newFolder().toPath();
     BuildImageSteps buildImageSteps =
         new BuildImageSteps(
             buildConfiguration,
             sourceFilesConfiguration,
-            temporaryCacheDirectory.getRoot().toPath());
+            Caches.newInitializer(cacheDirectory).setBaseCacheDirectory(cacheDirectory));
 
     long lastTime = System.nanoTime();
     buildImageSteps.run();