+++ jib-core/src/test/java/com/google/cloud/tools/jib/builder/TestBuildLogger.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Google Inc.
+ * Copyright 2018 Google LLC. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -42,5 +42,10 @@ public class TestBuildLogger implements BuildLogger {
   @Override
   public void error(CharSequence message) {
     LOGGER.error(message.toString());
+  }
+
+  @Override
+  public void lifecycle(CharSequence message) {
+    info(message);
   }
 }