+++ jib-core/src/test/java/com/google/cloud/tools/jib/registry/credentials/json/DockerConfigTemplateTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Google Inc.
+ * Copyright 2018 Google LLC. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -18,11 +18,8 @@ package com.google.cloud.tools.jib.registry.credentials.json;
 
 import com.google.cloud.tools.jib.json.JsonTemplateMapper;
 import com.google.common.io.Resources;
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.net.URISyntaxException;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import org.junit.Assert;
@@ -30,29 +27,6 @@ import org.junit.Test;
 
 /** Tests for {@link DockerConfigTemplate}. */
 public class DockerConfigTemplateTest {
-
-  @Test
-  public void test_toJson() throws URISyntaxException, IOException {
-    // Loads the expected JSON string.
-    Path jsonFile = Paths.get(Resources.getResource("json/dockerconfig.json").toURI());
-    String expectedJson = new String(Files.readAllBytes(jsonFile), StandardCharsets.UTF_8);
-
-    // Creates the JSON object to serialize.
-    DockerConfigTemplate dockerConfigTemplate =
-        new DockerConfigTemplate()
-            .addAuth("some registry", "some auth")
-            .addAuth("some other registry", "some other auth")
-            .addAuth("just registry", null)
-            .setCredsStore("some credential store")
-            .addCredHelper("some registry", "some credential helper")
-            .addCredHelper("another registry", "another credential helper");
-
-    // Serializes the JSON object.
-    ByteArrayOutputStream jsonStream = new ByteArrayOutputStream();
-    JsonTemplateMapper.toBlob(dockerConfigTemplate).writeTo(jsonStream);
-
-    Assert.assertEquals(expectedJson, jsonStream.toString());
-  }
 
   @Test
   public void test_fromJson() throws URISyntaxException, IOException {
@@ -74,6 +48,8 @@ public class DockerConfigTemplateTest {
         dockerConfigTemplate.getCredentialHelperFor("some other registry"));
     Assert.assertEquals(
         "some credential store", dockerConfigTemplate.getCredentialHelperFor("just registry"));
+    Assert.assertEquals(
+        "some credential store", dockerConfigTemplate.getCredentialHelperFor("with.protocol"));
     Assert.assertEquals(
         "another credential helper",
         dockerConfigTemplate.getCredentialHelperFor("another registry"));