+++ jib-gradle-plugin/src/main/java/com/google/cloud/tools/jib/gradle/ProjectProperties.java
@@ -17,7 +17,8 @@
 package com.google.cloud.tools.jib.gradle;
 
 import java.util.ArrayList;
-import org.gradle.api.Nullable;
+import java.util.List;
+import javax.annotation.Nullable;
 import org.gradle.api.Project;
 import org.gradle.api.Task;
 import org.gradle.api.logging.Logger;
@@ -37,7 +38,7 @@ class ProjectProperties {
   /** Extracts main class from 'jar' task, if available. */
   @Nullable
   String getMainClassFromJarTask() {
-    ArrayList<Task> jarTasks = new ArrayList<>(project.getTasksByName("jar", false));
+    List<Task> jarTasks = new ArrayList<>(project.getTasksByName("jar", false));
     if (jarTasks.size() != 1) {
       return null;
     }